Moved operations into ops package.
