+++ math/shared/src/main/scala/io/youi/math/prototype/ops/Matrix3Multiply.scala
@@ -1,4 +1,6 @@
-package io.youi.math.prototype
+package io.youi.math.prototype.ops
+
+import io.youi.math.prototype.Matrix3
 
 object Matrix3Multiply extends Matrix3Modifier {
   override def apply(m1: Matrix3, m2: Matrix3, modifier: Mod): Matrix3 = modifier(
